requirements.txt
openai>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
PyQt6>=6.4.0
python-dotenv>=1.0.0
tqdm>=4.65.0 